moved files one level up removing redundant folder
