+++ libs/platform/src/lib/platform.module.ts
@@ -26,7 +26,7 @@ import { PlatformThumbnailModule } from './components/thumbnail/public_api';
 import { PlatformObjectMarkerModule } from './components/object-marker/object-marker.module';
 import { PlatformObjectAttributeModule } from './components/object-attribute/object-attribute.module';
 import { PlatformTableModule } from './components/table/table.module';
-import { PlatformDatetimePickerModule } from './components/form/datetime-picker/datetime-picker/datetime-picker.module';
+import { PlatformDatetimePickerModule } from './components/form/datetime-picker/datetime-picker.module';
 
 @NgModule({
     imports: [CommonModule, FundamentalNgxCoreModule],
